pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Multi-document RAG pipeline with ChromaDB + Gemini"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "pypdf2>=3.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "gradio>=4.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
